\name{blvcm_bin_data}
\alias{blvcm_bin_data}
\docType{data}
\title{
Example data for BLVCM_bin
}
\description{
This is an example dataset consisting of binary traits for the blvcm_bin function.
}
\usage{data(blvcm_bin_data)}
\format{
  The format is:
List of 2
 $ pheno_data: num [1:2000, 1:3] 0 1 1 1 0 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:3] "pheno" "fam" "zyg"
 $ geno_data : int [1:2000, 1:40] 0 0 0 0 0 0 0 0 0 0 ...
}

\examples{
data(blvcm_bin_data)

}
\keyword{datasets}
