\name{blvcm_data}
\alias{blvcm_data}
\docType{data}
\title{
Example data for BLVCM
}
\description{
This is an example dataset consisting of continuous traits for the blvcm function.
}
\usage{data(blvcm_data)}
\format{
  The format is:
List of 2
 $ pheno_data: num [1:600, 1:3] -0.0813 -1.0135 0.4363 0.7927 0.9597 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : NULL
  .. ..$ : chr [1:3] "pheno" "fam" "zyg"
 $ geno_data : int [1:600, 1:40] 0 0 0 0 0 0 0 0 0 0 ...
}

\examples{
data(blvcm_data)

}
\keyword{datasets}
