\name{hbmr_bin_data}
\alias{hbmr_bin_data}
\docType{data}
\title{
Example data for hbmr_bin
}
\description{
This is an example dataset consisting of binary traits for the hbmr_bin function.
}
\usage{data(hbmr_bin_data)}
\format{
  The format is:
List of 2
 $ pheno: num [1:1720] 0 1 1 1 0 ...
 $ geno : int [1:1720, 1:40] 0 0 0 0 0 0 0 0 0 0 ...
 $ qi : int [1:1720, 1:40] 0 0 0 0 0 0 0 0 0 0 ...
 $ kin : int [1:1720, 1:1720] 0 0 0 0 0 0 0 0 0 0 ...
}

\examples{
data(hbmr_bin_data)
}