\name{hbmr_data}
\alias{hbmr_data}
\docType{data}
\title{
Example data for HBMR
}
\description{
This is an example dataset consisting of continuous traits for the hbmr function.
}
\usage{data(hbmr_data)}
\format{
  The format is:
List of 3
 $ pheno_data: num [1:600] -0.255 0.398 2.982 1.361 -0.165 ...
 $ geno_data : num [1:600, 1:50] 1 0 0 0 0 0 0 0 0 0 ...
 $ qual_data : num [1:600, 1:50] 5 5 5 99 99 99 99 99 99 99 ...
}

\examples{
data(hbmr_data)
}
\keyword{datasets}
